Add files via upload
A book shop databse for bookshop employees with functions that contain viewallBooks, void viewBook, searchBookTitle, searchBookAuthor, if its inStock, getBookPrice, addAbook, boughtSearchByTitleAndAuthor, boughtSearchByISBN, addingCopiesByTitleAndAuthor, addingCopiesByISBN, deleteBookByTitleAndAuthor, deleteBookByISBN. Goes through a file that is seperated by commas so a csv file. Some functions just traverse the file and compare user input to what is being looked for, and others store a line in a vector and compare a specific vector position to user input.